fix(pr-fleet-controller): split span folding below lint limit
Why
The repository verify gate rejects the dashboard fold module once its progress handling exceeds the max-lines rule.

What
Move span schemas and folding into a focused module, preserve the fold exports, and simplify failure-signal key handling.

Verification
cd packages/pr-fleet-controller/packages/web && bunx eslint src/lib/fold.ts src/lib/span-fold.ts
bun test packages/pr-fleet-controller/packages/web/src/lib/fold.test.ts
bunx turbo run typecheck --filter=@shepherdjerred/pr-fleet-web
