+++ packages/pr-fleet-controller/packages/web/src/lib/fold.ts
@@ -11,49 +11,25 @@ import {
   FleetFailureClassSchema,
   ProgressPayloadSchemas,
 } from "@shepherdjerred/pr-fleet-controller/src/progress-events.ts";
+import { applySpanLine as applySpanLineFromSpans } from "./span-fold.ts";
 
-const SpanRecordSchema = z.object({
-  id: z.string(),
-  traceId: z.string(),
-  parentSpanId: z.string().optional(),
-  name: z.string(),
-  type: z.string(),
-  entityName: z.string().optional(),
-  startTime: z.string().optional(),
-  endTime: z.string().optional(),
-  input: z.unknown().optional(),
-  output: z.unknown().optional(),
-  attributes: z.unknown().optional(),
-  metadata: z.record(z.string(), z.unknown()).optional(),
-  tags: z.array(z.string()).optional(),
-  errorInfo: z.unknown().optional(),
-  isRootSpan: z.boolean().optional(),
-});
-export type SpanRecord = z.infer<typeof SpanRecordSchema>;
-
-const LegacySpanLineSchema = z.discriminatedUnion("kind", [
-  z.object({ kind: z.literal("span"), span: SpanRecordSchema }),
-  z.object({ kind: z.literal("metric"), metric: z.unknown() }),
-]);
-
-const HrTimeSchema = z.tuple([z.number(), z.number()]);
-const OtelSpanSchema = z.object({
-  traceId: z.string(),
-  spanId: z.string(),
-  parentSpanId: z.string().optional(),
-  name: z.string(),
-  startTime: HrTimeSchema,
-  duration: HrTimeSchema,
-  status: z.object({ code: z.number(), message: z.string().optional() }),
-  attributes: z.record(z.string(), z.unknown()),
-  events: z.array(z.unknown()),
-});
-const OtelSpanLineSchema = z.object({
-  schemaVersion: z.literal(1),
-  kind: z.literal("span"),
-  span: OtelSpanSchema,
-});
-const SpanLineSchema = z.union([LegacySpanLineSchema, OtelSpanLineSchema]);
+export type SpanRecord = {
+  id: string;
+  traceId: string;
+  parentSpanId?: string | undefined;
+  name: string;
+  type: string;
+  entityName?: string | undefined;
+  startTime?: string | undefined;
+  endTime?: string | undefined;
+  input?: unknown;
+  output?: unknown;
+  attributes?: unknown;
+  metadata?: Record<string, unknown> | undefined;
+  tags?: string[] | undefined;
+  errorInfo?: unknown;
+  isRootSpan?: boolean | undefined;
+};
 
 export type TimelineItem =
   | { kind: "event"; t: number; order: number; event: RecordedRunEvent }
@@ -195,7 +171,7 @@ function recordFailure(
 
 function failureSignalKey(event: RecordedRunEvent): string | null {
   const toolCallId = event.correlation.toolCallId;
-  return toolCallId === undefined ? null : toolCallId;
+  return toolCallId ?? null;
 }
 
 function markFailureSignal(
@@ -480,76 +456,12 @@ function applyEvent(view: RunView, event: RecordedRunEvent): void {
   }
 }
 
-function spanPrNumber(view: RunView, span: SpanRecord): number | undefined {
-  const raw = span.metadata?.["prNumber"];
-  const parsed = z.number().int().positive().safeParse(raw);
-  return parsed.success ? parsed.data : view.tracePrNumbers.get(span.traceId);
-}
-
-function applySpan(view: RunView, span: SpanRecord): void {
-  const item: TimelineItem = {
-    kind: "span",
-    t: parsedEpoch(span.endTime ?? span.startTime),
-    order: view.counter++,
-    span,
-  };
-  const prNumber = spanPrNumber(view, span);
-  if (prNumber === undefined) {
-    if (view.fleetSpanIds.has(span.id)) {
-      return;
-    }
-    view.fleetSpanIds.add(span.id);
-    view.fleetTimeline.push(item);
-    return;
-  }
-  const pr = prView(view, prNumber);
-  if (pr.spanIds.has(span.id)) {
-    return;
-  }
-  pr.spanIds.add(span.id);
-  pr.timeline.push(item);
-}
-
 export function applyEventLine(view: RunView, raw: string): void {
   applyEvent(view, RecordedRunEventSchema.parse(JSON.parse(raw)));
 }
 
 export function applySpanLine(view: RunView, raw: string): void {
-  const line = SpanLineSchema.parse(JSON.parse(raw));
-  if (line.kind === "span") {
-    applySpan(
-      view,
-      "schemaVersion" in line ? normalizeOtelSpan(line.span) : line.span,
-    );
-  }
-  // Metric lines are captured for completeness but not rendered per-PR: metric
-  // labels intentionally omit trace/run ids, so they cannot be joined to a PR.
-  // Token/cost is surfaced from each model span's attributes instead.
-}
-
-function hrTimeMilliseconds(time: readonly [number, number]): number {
-  return time[0] * 1000 + time[1] / 1_000_000;
-}
-
-function normalizeOtelSpan(span: z.infer<typeof OtelSpanSchema>): SpanRecord {
-  const startMs = hrTimeMilliseconds(span.startTime);
-  const durationMs = hrTimeMilliseconds(span.duration);
-  return SpanRecordSchema.parse({
-    id: span.spanId,
-    traceId: span.traceId,
-    ...(span.parentSpanId === undefined
-      ? {}
-      : { parentSpanId: span.parentSpanId }),
-    name: span.name,
-    type:
-      span.name.includes("tool") || span.name.includes("Tool")
-        ? "tool_call"
-        : "model_inference",
-    startTime: new Date(startMs).toISOString(),
-    endTime: new Date(startMs + durationMs).toISOString(),
-    attributes: span.attributes,
-    output: { status: span.status, events: span.events },
-  });
+  applySpanLineFromSpans(view, raw);
 }
 
 /** Timeline sorted by event time, ties broken by arrival order. */
